+++ src/main/java/roomescape/business/domain/member/MemberName.java
@@ -1,4 +1,4 @@
-package roomescape.business.domain;
+package roomescape.business.domain.member;
 
 public record MemberName(
         String value
