+++ src/main/java/roomescape/business/domain/reservation/ReservationDateTimeFormatter.java
@@ -1,4 +1,4 @@
-package roomescape.business.domain;
+package roomescape.business.domain.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
