+++ src/main/java/roomescape/business/domain/reservation/ReservationTheme.java
@@ -1,4 +1,4 @@
-package roomescape.business.domain;
+package roomescape.business.domain.reservation;
 
 public final class ReservationTheme {
