+++ src/test/java/roomescape/business/domain/reservation/ReservationTimeTest.java
@@ -1,11 +1,10 @@
-package roomescape.business;
+package roomescape.business.domain.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.business.domain.ReservationTime;
 
 class ReservationTimeTest {
