+++ src/test/java/roomescape/business/service/ReservationThemeServiceTest.java
@@ -8,9 +8,9 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.business.domain.Reservation;
-import roomescape.business.domain.ReservationTheme;
-import roomescape.business.domain.ReservationTime;
+import roomescape.business.domain.reservation.Reservation;
+import roomescape.business.domain.reservation.ReservationTheme;
+import roomescape.business.domain.reservation.ReservationTime;
 import roomescape.exception.ReservationThemeException;
 import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.ReservationThemeRepository;
