+++ src/test/java/roomescape/persistence/fakerepository/FakeReservationTimeRepository.java
@@ -8,8 +8,8 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
-import roomescape.business.domain.ReservationDateTimeFormatter;
-import roomescape.business.domain.ReservationTime;
+import roomescape.business.domain.reservation.ReservationDateTimeFormatter;
+import roomescape.business.domain.reservation.ReservationTime;
 import roomescape.persistence.ReservationTimeRepository;
 import roomescape.persistence.entity.ReservationTimeEntity;
 import roomescape.presentation.dto.AvailableTimesResponseDto;
